+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # 1.x branch
 ## 1.7 branch
 ### 1.7.4
-* `addSlashTerm()`, `getExtension()`, `getRoot()`, `normalizePath()` method provided by `Filesystem` class can now be
-    called statically.
+* all methods provided by the `Filesystem` class (except for `makePathRelative()`) can now be called statically.
 
 ### 1.7.3
 * fixed a bug for `Exceptionist::isInstanceOf()`: this method now accepts instantiated objects or class name as string